requirements.txt
streamlit==1.32.0
spacy==3.7.2
python-docx==1.1.0
pytest==8.0.0
pytest-cov==4.1.0
python-dotenv==1.0.1
rapidfuzz>=2.13.7
flake8==7.0.0
mypy==1.8.0
jinja2==3.1.3
click==8.1.7
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
black==24.1.1
safety==2.3.5
packaging<24,>=23